pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-memex"
dynamic = ["version"]
description = "Personal conversation knowledge base: import, search, and analyze conversations from ChatGPT, Claude, Gemini, and Claude Code"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Alex Towell", email = "lex@metafunctor.com"},
]
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastmcp<3",
    "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "pytest-asyncio"]

[project.urls]
Homepage = "https://github.com/queelius/llm-memex"
Repository = "https://github.com/queelius/llm-memex"
Issues = "https://github.com/queelius/llm-memex/issues"

[project.scripts]
llm-memex = "llm_memex.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests/llm_memex"]
asyncio_mode = "auto"

[tool.setuptools.dynamic]
version = {attr = "llm_memex.__version__"}

[tool.setuptools.packages.find]
include = ["llm_memex*"]

[tool.setuptools.package-data]
"llm_memex.scripts" = ["patterns/*.txt"]
"llm_memex.exporters" = ["vendored/*"]
